requirements.txt
prometheus-client>=0.16.0
pyyaml>=6.0
requests>=2.28.2
jsonpath-ng>=1.5.0
